ChessMod3coordinatesBound
I want to be able to add coordinates that are outside of the chess
board. So I will rely on a check I created in Piece instead of the
Coordinate. By doing this I will be able to generate possible moves
better, and I will remove a few "magic numbers"